Implement dedicated health check extension methods for improved configuration and management
